+++ CHANGELOG.md
@@ -7,6 +7,8 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.10.1] - 2025-10-29
+
 ### Fixed
 
 - Prevent deadlocks when dropping instances of `AsyncClient` that have not been disconnected before
@@ -553,7 +555,8 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - First public release.
 
-[Unreleased]: https://github.com/HMIProject/open62541/compare/v0.10.0...HEAD
+[Unreleased]: https://github.com/HMIProject/open62541/compare/v0.10.1...HEAD
+[0.10.1]: https://github.com/HMIProject/open62541/compare/v0.10.0...v0.10.1
 [0.10.0]: https://github.com/HMIProject/open62541/compare/v0.9.4...v0.10.0
 [0.9.4]: https://github.com/HMIProject/open62541/compare/v0.9.3...v0.9.4
 [0.9.3]: https://github.com/HMIProject/open62541/compare/v0.9.2...v0.9.3